requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
pyjwt
python-dotenv
redis
bcrypt